doc: Organized front and back-end directories and rewrote the bat script.
