--- app/tools/config.json
@@ -1,5 +1,0 @@
-[
-  "Easyfile用况图文档(1)",
-  {"str":"用户","typeface":"","size":0,"color":""},
-  {"str":"我","typeface":"宋体","size":12,"color":"66ccff"}
-]
