+++ backend/app/tools/myfont.py
@@ -1,0 +1,15 @@
+class Font:
+    def __init__(self, color="", name="", size=0):
+        self.color = color
+        self.name = name
+        self.size = size
+
+
+class RepStr:
+    def __init__(self, str="", color="", name="", size=0):
+        self.str = str
+        self.font = Font(color, name, size)
+
+
+    def from_font(self,str="",font=Font()):
+        return RepStr(str,font.color,font.name,font.size)
