--- instance/.gitignore
@@ -1,4 +1,0 @@
-*
-
-!.gitignore
-!config_sample.py
